fixed symbol duplications caused by the bad coding style of in the given filesystem...

+++ data_storage/type_tag.h
@@ -9,7 +9,7 @@
 
 namespace watery {
 
-enum struct TypeTag : uint32_t {
+enum struct TypeTag : uint16_t {
     INTEGER
 };
